+++ pysaliency/precomputed_models.py
@@ -2,7 +2,9 @@ from __future__ import print_function, division, absolute_import
 
 import glob
 import os.path
+import tarfile
 import warnings
+import zipfile
 
 import numpy as np
 from imageio import imread
@@ -205,13 +207,133 @@ class HDF5Model(Model):
     """
     def __init__(self, stimuli, filename, check_shape=True, **kwargs):
         super(HDF5Model, self).__init__(**kwargs)
-        self.parent_model = HDF5SaliencyMapModel(stimuli = stimuli,
-                                                 filename = filename,
-                                                 caching=False,
-                                                 check_shape=check_shape)
+        self.parent_model = HDF5SaliencyMapModel(
+            stimuli=stimuli,
+            filename=filename,
+            caching=False,
+            check_shape=check_shape
+        )
 
     def _log_density(self, stimulus):
         smap = self.parent_model.saliency_map(stimulus)
         if not -0.01 <= logsumexp(smap) <= 0.01:
             raise ValueError('Not a correct log density!')
         return smap
+
+
+class TarFileLikeZipFile(object):
+    """ Wrapper that makes TarFile behave more like ZipFile """
+    def __init__(self, filename, *args, **kwargs):
+        self.tarfile = tarfile.open(filename, *args, **kwargs)
+    def namelist(self):
+        filenames = []
+        for tar_info in self.tarfile.getmembers():
+            filenames.append(tar_info.name)
+        return filenames
+
+    def open(self, name, mode='r'):
+        return self.tarfile.extractfile(name)
+
+
+class PredictionsFromArchiveMixin(object):
+    def __init__(self, stimuli, archive_file, *args, **kwargs):
+        if not isinstance(stimuli, FileStimuli):
+            raise TypeError('PredictionsFromArchiveMixin works only with FileStimuli!')
+
+        super(PredictionsFromArchiveMixin, self).__init__(*args, **kwargs)
+
+        self.stimuli = stimuli
+        self.stimulus_ids = list(stimuli.stimulus_ids)
+        self.archive_file = archive_file
+        _, archive_ext = os.path.splitext(self.archive_file)
+        if archive_ext.lower() == '.zip':
+            self.archive = zipfile.ZipFile(self.archive_file)
+        elif archive_ext.lower() == '.tar':
+            self.archive = TarFileLikeZipFile(self.archive_file)
+        elif archive_ext.lower() == '.gz':
+            self.archive = TarFileLikeZipFile(self.archive_file)
+        elif archive_ext.lower() == '.rar':
+            import rarfile
+            self.archive = rarfile.RarFile(self.archive_file)
+        else:
+            raise ValueError(archive_file)
+
+        files = self.archive.namelist()
+        files = [f for f in files if '.ds_store' not in f.lower()]
+        files = [f for f in files if '__macosx' not in f.lower()]
+        stems = [os.path.splitext(f)[0] for f in files]
+
+        stimuli_files = get_minimal_unique_filenames(stimuli.filenames)
+        stimuli_stems = [os.path.splitext(f)[0] for f in stimuli_files]
+
+        prediction_filenames = []
+        for stimuli_stem in stimuli_stems:
+            candidates = [stem for stem in stems if stem.endswith(stimuli_stem)]
+            if not candidates:
+                raise ValueError("Can't find file for {}".format(stimuli_stem))
+            if len(candidates) > 1:
+                raise ValueError("Multiple candidates for {}: {}", stimuli_stem, candidates)
+
+            target_stem, = candidates
+            target_index = stems.index(target_stem)
+            target_filename = files[target_index]
+
+            prediction_filenames.append(target_filename)
+
+        self.files = prediction_filenames
+
+    def _prediction(self, stimulus):
+        stimulus_id = get_image_hash(stimulus)
+        stimulus_index = self.stimuli.stimulus_ids.index(stimulus_id)
+        filename = self.files[stimulus_index]
+        return self._load_file(filename)
+
+    def _load_file(self, filename):
+        _, ext = os.path.splitext(filename)
+
+        content = self.archive.open(filename)
+
+        if ext.lower() in ['.png', '.jpg', '.jpeg', '.tiff']:
+            return imread(content).astype(float)
+        elif ext.lower() == '.npy':
+            return np.load(content).astype(float)
+        elif ext.lower() == '.mat':
+            data = loadmat(content)
+            variables = [v for v in data if not v.startswith('__')]
+            if len(variables) > 1:
+                raise ValueError('{} contains more than one variable: {}'.format(filename, variables))
+            elif len(variables) == 0:
+                raise ValueError('{} contains no data'.format(filename))
+            return data[variables[0]]
+        else:
+            raise ValueError('Unkown file type: {}'.format(ext))
+
+    @staticmethod
+    def can_handle(filename):
+        try:
+            import rarfile
+            return zipfile.is_zipfile(filename) or tarfile.is_tarfile(filename) or rarfile.is_rarfile(filename)
+        except ImportError:
+            warnings.warn("Can't test for rarfiles since rarfile is not installed!")
+            return zipfile.is_zipfile(filename) or tarfile.is_tarfile(filename)
+
+
+class SaliencyMapModelFromArchive(PredictionsFromArchiveMixin, SaliencyMapModel):
+    def __init__(self, stimuli, archive_file, **kwargs):
+        super(SaliencyMapModelFromArchive, self).__init__(stimuli, archive_file, **kwargs)
+
+    def _saliency_map(self, stimulus):
+        return self._prediction(stimulus)
+
+
+class ModelFromArchive(PredictionsFromArchiveMixin, Model):
+    def __init__(self, stimuli, archive_file, **kwargs):
+        super(ModelFromArchive, self).__init__(stimuli, archive_file, **kwargs)
+
+    def _log_density(self, stimulus):
+        logdensity = self._prediction(stimulus)
+        density_sum = logsumexp(logdensity)
+        if np.abs(density_sum) > 0.01:
+            logdensity_sum = np.sum(logdensity)
+            raise ValueError("predictions not normalized: log of sum is {}, sum is {}".format(density_sum, logdensity_sum))
+        return logdensity